refactor: move table logic to separated folder
