--- Source/SeaInk.Core/Models/Google/Exceptions/DriveException.cs
@@ -1,13 +1,0 @@
-using System;
-
-namespace SeaInk.Core.Models.Google.Exceptions
-{
-    public class DriveException : Exception
-    {
-        public DriveException(string message)
-            : base(message) { }
-
-        public DriveException()
-            : base("Unspecified drive exception") { }
-    }
-}
