--- Source/SeaInk.Core/Models/Google/GoogleFileTypes.cs
@@ -1,8 +1,0 @@
-namespace SeaInk.Core.Models.Google
-{
-    public static class GoogleFileTypes
-    {
-        public static readonly string Folder = @"application/vnd.google-apps.folder";
-        public static readonly string GoogleSpreadsheet = @"application/vnd.google-apps.spreadsheet";
-    }
-}
