--- Source/SeaInk.Core/Models/Tables/Enums/FontStyle.cs
@@ -1,11 +1,0 @@
-namespace SeaInk.Core.Models.Tables.Enums
-{
-    public enum FontStyle
-    {
-        Regular,
-        Bold,
-        Italic,
-        Crossed,
-        Underlined
-    }
-}
