--- Source/SeaInk.Core/Models/Tables/Exceptions/ExistingSheetException.cs
@@ -1,15 +1,0 @@
-namespace SeaInk.Core.Models.Tables.Exceptions
-{
-    public class ExistingSheetException : TableException
-    {
-        public ExistingSheetException(string message) 
-            : base($"Sheet with specified signature already exists {message}")
-        {
-        }
-
-        public ExistingSheetException()
-            : this("")
-        {
-        }
-    }
-}
