--- Source/SeaInk.Core/Models/Tables/Exceptions/InvalidRangeBoundsException.cs
@@ -1,15 +1,0 @@
-namespace SeaInk.Core.Models.Tables.Exceptions
-{
-    public class InvalidRangeBoundsException : TableException
-    {
-        public InvalidRangeBoundsException(string message)
-            : base($"Cannot create range with given bounds {message}")
-        {
-        }
-
-        public InvalidRangeBoundsException()
-            : this("")
-        {
-        }
-    }
-}
