--- Source/SeaInk.Core/Models/Tables/Exceptions/NonExistingIndexException.cs
@@ -1,15 +1,0 @@
-namespace SeaInk.Core.Models.Tables.Exceptions
-{
-    public class NonExistingIndexException : TableException
-    {
-        public NonExistingIndexException(string message)
-            : base($"Specified index does not exists {message}")
-        {
-        }
-
-        public NonExistingIndexException()
-            : this("")
-        {
-        }
-    }
-}
