+++ Source/SeaInk.Core/TableIntegrations/Google/Exceptions/DriveException.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace SeaInk.Core.TableIntegrations.Google.Exceptions
+{
+    public class DriveException : Exception
+    {
+        public DriveException(string message)
+            : base(message) { }
+
+        public DriveException()
+            : base("Unspecified drive exception") { }
+    }
+}
