+++ Source/SeaInk.Core/TableIntegrations/Google/Exceptions/InsufficientPermissionException.cs
@@ -1,0 +1,11 @@
+namespace SeaInk.Core.TableIntegrations.Google.Exceptions
+{
+    public class InsufficientPermissionException : DriveException
+    {
+        public InsufficientPermissionException(string message)
+            : base($"Insufficient permission for accessing google drive\n{message}") { }
+
+        public InsufficientPermissionException()
+            : this("") { }
+    }
+}
