+++ Source/SeaInk.Core/TableIntegrations/Google/GoogleFileTypes.cs
@@ -1,0 +1,8 @@
+namespace SeaInk.Core.TableIntegrations.Google
+{
+    public static class GoogleFileTypes
+    {
+        public static readonly string Folder = @"application/vnd.google-apps.folder";
+        public static readonly string GoogleSpreadsheet = @"application/vnd.google-apps.spreadsheet";
+    }
+}
