+++ Source/SeaInk.Core/TableIntegrations/Models/Exceptions/TableNotLoadedException.cs
@@ -1,0 +1,15 @@
+namespace SeaInk.Core.TableIntegrations.Models.Exceptions
+{
+    public class TableNotLoadedException : TableException
+    {
+        public TableNotLoadedException(string message)
+            : base($"Table is not loaded {message}")
+        {
+        }
+
+        public TableNotLoadedException()
+            : this("")
+        {
+        }
+    }
+}
