+++ Source/SeaInk.Core/TableIntegrations/Models/Styles/Enums/FontStyle.cs
@@ -1,0 +1,11 @@
+namespace SeaInk.Core.TableIntegrations.Models.Styles.Enums
+{
+    public enum FontStyle
+    {
+        Regular,
+        Bold,
+        Italic,
+        Crossed,
+        Underlined
+    }
+}
